refactor: remove unnecessary reexports
