+++ src/index.ts
@@ -22,10 +22,12 @@ export * from './utils/bytes';
 export * from './tx';
 export * from './tx/builder';
 export * from './tx/builder/helpers';
+export * from './tx/builder/constants';
 export * from './tx/builder/schema';
 export * from './tx/builder/address';
 export * from './utils/amount-formatter';
 export * from './utils/hd-wallet';
+export { sha256hash, encode, decode } from './utils/encoder';
 export * from './aens';
 export * from './contract/methods';
 export * from './oracle';
